@@ -1,9 +1,9 @@
 // @ts-check
 // Note: type annotations allow type checking and IDEs autocompletion
+import { themes } from "prism-react-renderer";
+import { themes as prismThemes } from "prism-react-renderer";
 
-const lightCodeTheme = require("prism-react-renderer/themes/github");
-const darkCodeTheme = require("prism-react-renderer/themes/dracula");
-
+const orgName = "open-tech-foundation";
 const repoName = "";
 const pkgVer = ""; //require(`../packages/${repoName}/package.json`).version;
 
@@ -13,10 +13,11 @@ const config = {
   tagline: "Tagline.",
   url: "https://your-docusaurus-test-site.com",
   baseUrl: "/",
+  trailingSlash: false,
   onBrokenLinks: "throw",
   onBrokenMarkdownLinks: "warn",
-  favicon: "img/favicon.ico",
-  organizationName: "open-tech-world", // Usually your GitHub org/user name.
+  favicon: "img/Logo.svg",
+  organizationName: orgName, // Usually your GitHub org/user name.
   projectName: repoName, // Usually your repo name.
 
   presets: [
@@ -28,7 +29,7 @@ const config = {
           sidebarPath: require.resolve("./sidebars.js"),
           sidebarCollapsed: false,
           // Please change this to your repo.
-          editUrl: `https://github.com/open-tech-world/${repoName}/tree/main/`,
+          editUrl: `https://github.com/${orgName}/${repoName}/tree/main/`,
         },
         blog: false,
         theme: {
@@ -44,23 +45,24 @@ const config = {
       navbar: {
         title: "My Site",
         logo: {
-          alt: "Open Tech World Logo",
-          src: "img/logo.png",
+          alt: "Logo",
+          src: "img/Logo.svg",
         },
         items: [
           {
-            label: "v" + pkgVer,
+            html: '<img src="/img/Npm-logo.svg" width="48" height="32" />',
             position: "right",
-            href: `https://www.npmjs.com/package/@open-tech-world/${repoName}`,
+            href: `https://www.npmjs.com/package/@opentf/${repoName}`,
+            className: "npm-link"
           },
           {
-            type: "doc",
-            docId: "Getting Started",
+            type: "docSidebar",
+            sidebarId: 'docsSidebar',
             position: "left",
             label: "Docs",
           },
           {
-            href: `https://github.com/open-tech-world/${repoName}`,
+            href: `https://github.com/${orgName}/${repoName}`,
             className: "header-github-link",
             position: "right",
           },
@@ -70,17 +72,14 @@ const config = {
         style: "dark",
         links: [
           {
-            html: `⚡ by <a href="https://open-tech-world.pages.dev">Open Tech World</a>`,
-          },
-          {
-            html: `📝 with <a href="https://docusaurus.io/">Docusaurus</a>`,
+            html: `<a href="https://docusaurus.io/">📝 Built with Docusaurus.</a>`,
           },
         ],
-        copyright: `Copyright © ${new Date().getFullYear()} Contributors of <a href="https://github.com/open-tech-world/${repoName}">@open-tech-world/${repoName}</a>`,
+        copyright: `Copyright © ${new Date().getFullYear()} <a href="https://open-tech-foundation.pages.dev">Open Tech Foundation</a>.`,
       },
       prism: {
-        theme: lightCodeTheme,
-        darkTheme: darkCodeTheme,
+        theme: themes.github,
+        darkTheme: themes.dracula,
       },
     }),
 };
