@@ -4,6 +4,9 @@ import Layout from "@theme/Layout";
 import Link from "@docusaurus/Link";
 import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
 import styles from "./index.module.css";
+import HeadingWithDivider from "@site/src/components/HeadingWithDivider";
+import { FcCheckmark } from "react-icons/fc";
+import { FcRight } from "react-icons/fc";
 
 function HomepageHeader() {
   const { siteConfig } = useDocusaurusContext();
@@ -15,9 +18,11 @@ function HomepageHeader() {
         <div className={styles.buttons}>
           <Link
             className="button button--secondary button--lg"
-            to="/docs/Getting Started"
+            to="/docs/getting-started"
+            style={{ display: "flex", justifyContent: "center" }}
           >
-            Getting Started 🚀
+            Getting Started{" "}
+            <FcRight style={{ fontSize: "26px", marginLeft: "8px" }} />
           </Link>
         </div>
       </div>
@@ -33,13 +38,17 @@ export default function Home() {
       description="Description will go into a meta tag in <head />"
     >
       <HomepageHeader />
-      <main>
-      <div className={styles.features}>
-          <div className={styles.featuresText}>
-            <h3>FEATURES</h3>
-          </div>
+      <main style={{ padding: "25px" }}>
+        <HeadingWithDivider title="FEATURES" />
+        <div className={styles.features}>
           <ul>
-            <li> ✔️ Simple API</li>
+            <li>
+              {" "}
+              <FcCheckmark
+                style={{ fontSize: "24px", marginRight: "8px" }}
+              />{" "}
+              Simple API
+            </li>
           </ul>
         </div>
       </main>
